refactor: 重构 markdown 渲染器，分离数学渲染逻辑并优化代码结构
