+++ webclipper/src/entrypoints/popup/main.tsx
@@ -4,7 +4,6 @@ import PopupShell from '@ui/popup/PopupShell';
 import '@ui/styles/tokens.css';
 import '@ui/styles/buttons.css';
 import '@ui/styles/tailwind.css';
-import 'katex/dist/katex.min.css';
 import '@entrypoints/popup/style.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
